# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: AuthUserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  employeeCode?: string | null;
  hasFaceRegistered?: boolean;
}

export type EmployeeRole = typeof EmployeeRole[keyof typeof EmployeeRole];


export const EmployeeRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface Employee {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  employeeCode?: string | null;
  /** @nullable */
  department?: string | null;
  role: EmployeeRole;
  isActive: boolean;
  hasFaceRegistered: boolean;
  createdAt: string;
}

export type EmployeeInputRole = typeof EmployeeInputRole[keyof typeof EmployeeInputRole];


export const EmployeeInputRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface EmployeeInput {
  name: string;
  email: string;
  password: string;
  employeeCode?: string;
  department?: string;
  role: EmployeeInputRole;
}

export type EmployeeUpdateRole = typeof EmployeeUpdateRole[keyof typeof EmployeeUpdateRole];


export const EmployeeUpdateRole = {
  admin: 'admin',
  employee: 'employee',
} as const;

export interface EmployeeUpdate {
  name?: string;
  email?: string;
  employeeCode?: string;
  department?: string;
  role?: EmployeeUpdateRole;
  isActive?: boolean;
  password?: string;
}

export interface FaceDescriptorInput {
  /** Array of face descriptor arrays (each 128-element Float32Array serialized as number[]) */
  descriptors: number[][];
}

export interface Location {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  latitude: number;
  longitude: number;
  /** Geofence radius in meters */
  radius: number;
  isActive: boolean;
  createdAt: string;
}

export interface LocationInput {
  name: string;
  address?: string;
  latitude: number;
  longitude: number;
  /** Geofence radius in meters (default 100) */
  radius: number;
}

export interface LocationUpdate {
  name?: string;
  address?: string;
  latitude?: number;
  longitude?: number;
  /** Geofence radius in meters */
  radius?: number;
  isActive?: boolean;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  late: 'late',
  absent: 'absent',
} as const;

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  employeeName?: string;
  /** @nullable */
  employeeCode?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  locationName?: string | null;
  date: string;
  /** @nullable */
  checkInTime?: string | null;
  /** @nullable */
  checkOutTime?: string | null;
  status: AttendanceRecordStatus;
  faceVerified: boolean;
  locationVerified: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CheckInInput {
  employeeId: number;
  locationId: number;
  latitude: number;
  longitude: number;
  /** Single face descriptor (128 numbers) */
  faceDescriptor: number[];
}

export interface CheckOutInput {
  attendanceId: number;
  latitude: number;
  longitude: number;
}

export type AttendanceUpdateStatus = typeof AttendanceUpdateStatus[keyof typeof AttendanceUpdateStatus];


export const AttendanceUpdateStatus = {
  present: 'present',
  late: 'late',
  absent: 'absent',
} as const;

export interface AttendanceUpdate {
  status?: AttendanceUpdateStatus;
  checkInTime?: string;
  checkOutTime?: string;
  notes?: string;
}

export interface DashboardSummary {
  totalEmployees: number;
  presentToday: number;
  absentToday: number;
  lateToday: number;
  /** Percentage 0-100 */
  attendanceRate: number;
  checkedOutToday?: number;
  activeLocations?: number;
}

export interface TodayStats {
  date: string;
  records: AttendanceRecord[];
}

export interface DepartmentStat {
  department: string;
  total: number;
  present: number;
  absent: number;
  late: number;
  attendanceRate: number;
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  'check-in': 'check-in',
  'check-out': 'check-out',
} as const;

export type ActivityEventStatus = typeof ActivityEventStatus[keyof typeof ActivityEventStatus];


export const ActivityEventStatus = {
  present: 'present',
  late: 'late',
  absent: 'absent',
} as const;

export interface ActivityEvent {
  id: number;
  type: ActivityEventType;
  employeeName: string;
  /** @nullable */
  employeeCode?: string | null;
  /** @nullable */
  department?: string | null;
  locationName: string;
  timestamp: string;
  status?: ActivityEventStatus;
}

export type ListEmployeesParams = {
search?: string;
department?: string;
isActive?: boolean;
};

export type ListAttendanceParams = {
employeeId?: number;
locationId?: number;
date?: string;
startDate?: string;
endDate?: string;
status?: ListAttendanceStatus;
};

export type ListAttendanceStatus = typeof ListAttendanceStatus[keyof typeof ListAttendanceStatus];


export const ListAttendanceStatus = {
  present: 'present',
  late: 'late',
  absent: 'absent',
} as const;