# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Attendance Software API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Employees ──────────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: locationId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [present, late, absent]
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Mark attendance check-in (face + geofence verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "201":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Face not verified or outside geofence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Mark attendance check-out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckOutInput"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/{id}:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get a single attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Manually update an attendance record (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, employee]
        department:
          type: ["string", "null"]
        employeeCode:
          type: ["string", "null"]
        hasFaceRegistered:
          type: boolean

    # Employees
...
    AttendanceRecord:
      type: object
      required: [id, employeeId, date, status, faceVerified, locationVerified, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeCode:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        date:
          type: string
          format: date
        checkInTime:
          type: ["string", "null"]
          format: date-time
        checkOutTime:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [present, late, absent]
        faceVerified:
          type: boolean
        locationVerified:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CheckInInput:
      type: object
      required: [employeeId, locationId, latitude, longitude, faceDescriptor]
      properties:
        employeeId:
          type: integer
        locationId:
          type: integer
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        faceDescriptor:
          type: array
          description: Single face descriptor (128 numbers)
          items:
            type: number
    CheckOutInput:
      type: object
      required: [attendanceId, latitude, longitude]
      properties:
        attendanceId:
          type: integer
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
    AttendanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [present, late, absent]
        checkInTime:
          type: string
          format: date-time
        checkOutTime:
          type: string
          format: date-time
        notes:
          type: string

    # Dashboard
...