# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: AuthUserRole;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  employeeCode?: string | null;
  hasFaceRegistered?: boolean;
}