# lib/api-zod/src/generated/types/checkInInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInInput {
  employeeId: number;
  locationId: number;
  latitude: number;
  longitude: number;
  /** Single face descriptor (128 numbers) */
  faceDescriptor: number[];
}