# lib/api-zod/src/generated/types/checkOutInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Attendance Software API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckOutInput {
  attendanceId: number;
  latitude: number;
  longitude: number;
}